+++ yora-parser/src/main/java/dev/yekllurt/parser/interpreter/nativ/function/impl/SinNativeFunction.java
@@ -13,7 +13,7 @@ public class SinNativeFunction implements NativeFunction {
         if (Objects.isNull(parameters) || parameters.length != 1) {
             throw new ExecutionError("The sin function only accepts one parameter");
         }
-        return Optional.of((float) Math.sin(Float.parseFloat(String.valueOf(parameters[0]))));
+        return Optional.of(Math.sin(Double.parseDouble(String.valueOf(parameters[0]))));
     }
 
 }
